rework test_helpers
