+++ common/file.go
@@ -4,13 +4,54 @@ import (
 	"fmt"
 	"io"
 	"os"
+	"path"
+	"path/filepath"
+	"strings"
 )
 
-// Copied from https://stackoverflow.com/questions/21060945/simple-way-to-copy-a-file-in-golang
+// CopyDir will walk through each file in srcDir, and call
+// CopyFile on that file into dstDir.
+func CopyDir(srcDir string, dstDir string) (err error) {
+	cropNDirs := len(strings.Split(srcDir, "/"))
+	walkFn := func(iFilepath string, info os.FileInfo, err error) error {
+
+		// swap out the directory 'content_sanctuary' with 'content'
+		inDirs := strings.Split(iFilepath, "/")
+		outDirs := append([]string{dstDir}, inDirs[cropNDirs:]...)
+
+		dstFilepath := path.Join(outDirs...)
+
+		if info.Mode().IsDir() {
+			os.MkdirAll(dstFilepath, 0755)
+		} else {
+
+			if !info.Mode().IsRegular() {
+				return nil
+			}
+
+			if err != nil {
+				return err
+			}
+
+			err = CopyFile(iFilepath, dstFilepath)
+			if err != nil {
+				return err
+			}
+		}
+		return nil
+	}
+
+	err = filepath.Walk(srcDir, walkFn)
+	if err != nil {
+		return err
+	}
+	return nil
+}
+
+// Copied and modified from https://stackoverflow.com/questions/21060945/simple-way-to-copy-a-file-in-golang
 // CopyFile copies a file from src to dst. If src and dst files exist, and are
-// the same, then return success. Otherise, attempt to create a hard link
-// between the two files. If that fail, copy the file contents from src to dst.
-func CopyFile(src, dst string) (err error) {
+// the same, then return success. Copy the file contents from src to dst.
+func CopyFile(src string, dst string) (err error) {
 	sfi, err := os.Stat(src)
 	if err != nil {
 		return
@@ -29,11 +70,9 @@ func CopyFile(src, dst string) (err error) {
 		if !(dfi.Mode().IsRegular()) {
 			return fmt.Errorf("CopyFile: non-regular destination file %s (%q)", dfi.Name(), dfi.Mode().String())
 		}
-		if os.SameFile(sfi, dfi) {
-			return
-		}
+
 	}
-	if err = os.Link(src, dst); err == nil {
+	if os.SameFile(sfi, dfi) {
 		return
 	}
 	err = copyFileContents(src, dst)
